requirements.txt
numpy>=1.26
scipy>=1.11
matplotlib>=3.6
pandas>=2.1
Pillow>=10.0
PyYAML>=6.0

